+++ R/gh-package.R
@@ -1,0 +1,14 @@
+#' @keywords internal
+#' @aliases gh-package
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+#' @importFrom httr content add_headers headers status_code http_type GET POST
+#'   PATCH PUT DELETE
+#' @importFrom jsonlite fromJSON toJSON
+#' @importFrom utils URLencode capture.output
+#' @importFrom cli cli_status cli_status_update
+## usethis namespace: end
+NULL