+++ R/gh_rate_limit.R
@@ -4,8 +4,8 @@
 #' either pulls this information from a previous successful request
 #' or directly from the GitHub API.
 #'
-#' Further details on GitHub's API rate limit policies are available
-#' [here](https://developer.github.com/v3/#rate-limiting)
+#' Further details on GitHub's API rate limit policies are available at
+#' <https://docs.github.com/v3/#rate-limiting>.
 #'
 #' @param response `gh_response` object from a previous `gh` call, rate
 #' limit values are determined from values in the response header.
@@ -18,12 +18,14 @@
 #'
 #' @export
 
-gh_rate_limit = function(response = NULL, .token = NULL, .api_url = NULL, .send_headers = NULL) {
+gh_rate_limit <- function(response = NULL, .token = NULL, .api_url = NULL, .send_headers = NULL) {
   if (is.null(response)) {
     # This end point does not count against limit
     .token <- .token %||% gh_token(.api_url)
-    response <- gh("GET /rate_limit", .token = .token,
-                  .api_url = .api_url, .send_headers = .send_headers)
+    response <- gh("GET /rate_limit",
+      .token = .token,
+      .api_url = .api_url, .send_headers = .send_headers
+    )
   }
 
   stopifnot(inherits(response, "gh_response"))