+++ tests/testthat/test-errors.R
@@ -1,10 +1,8 @@
-context("github_error")
-
 test_that("errors return a github_error object", {
-
-  skip_if_offline()
+  skip_if_offline("github.com")
   skip_on_cran()
-  
+  skip_if_no_token()
+
   e <- tryCatch(gh("/missing", .token = tt()), error = identity)
 
   expect_s3_class(e, "github_error")
@@ -12,10 +10,10 @@ test_that("errors return a github_error object", {
 })
 
 test_that("can catch a given status directly", {
-
-  skip_if_offline()
+  skip_if_offline("github.com")
   skip_on_cran()
-  
+  skip_if_no_token()
+
   e <- tryCatch(
     gh("/missing", .token = tt()),
     "http_error_404" = identity