+++ tests/testthat/test-old-templates.R
@@ -1,0 +1,5 @@
+TMPL <- function(x) {
+  gsub("[{]([^}]+)[}]", ":\\1", x)
+}
+
+source("test-mock-repos.R", local = TRUE)